refactor(cuga-lite): thin cuga_lite_graph.py — move all logic into AgentGraphAdapter
cuga_lite_graph.py goes from ~1800 lines to 214: state class, loop adapter,
two thin wrappers, and the wiring factory. All node logic (prepare, sandbox,
bind-tools helpers, find-tools helpers, todos, knowledge, reflection) now
lives in AgentGraphAdapter, matching the pattern already used by
cuga_supervisor_graph.py / SupervisorGraphAdapter.

Also fixes 7 CodeRabbit review items from PR #250:
- graph_nodes.py: use model_copy(update=...) on HumanMessage to preserve id/additional_kwargs
- shared_nodes.py: same model_copy fix; Loguru %d/%s placeholders -> {}
- tool_approval_handler.py: handle_denial clears user_approved from metadata in state update
- cuga_supervisor_graph.py: remove dead _resolve_names_from_caller_frame + inspect import
- supervisor_graph_adapter.py: guard var_manager None; fix create_update_todos_tool import source
